+++ xfrm_policy_linux.go
@@ -40,7 +40,7 @@ func XfrmPolicyAdd(policy *XfrmPolicy) error {
 // XfrmPolicyAdd will add an xfrm policy to the system.
 // Equivalent to: `ip xfrm policy add $policy`
 func (h *Handle) XfrmPolicyAdd(policy *XfrmPolicy) error {
-	return h.xfrmPolicyAddOrUpdate(policy, nl.XFRM_MSG_NEWPOLICY)
+	return h.xfrmPolicyAddOrUpdate(policy, int(nl.XFRM_MSG_NEWPOLICY))
 }
 
 // XfrmPolicyUpdate will update an xfrm policy to the system.
@@ -52,7 +52,7 @@ func XfrmPolicyUpdate(policy *XfrmPolicy) error {
 // XfrmPolicyUpdate will update an xfrm policy to the system.
 // Equivalent to: `ip xfrm policy update $policy`
 func (h *Handle) XfrmPolicyUpdate(policy *XfrmPolicy) error {
-	return h.xfrmPolicyAddOrUpdate(policy, nl.XFRM_MSG_UPDPOLICY)
+	return h.xfrmPolicyAddOrUpdate(policy, int(nl.XFRM_MSG_UPDPOLICY))
 }
 
 func (h *Handle) xfrmPolicyAddOrUpdate(policy *XfrmPolicy, nlProto int) error {
@@ -115,7 +115,7 @@ func XfrmPolicyDel(policy *XfrmPolicy) error {
 // the Tmpls are ignored when matching the policy to delete.
 // Equivalent to: `ip xfrm policy del $policy`
 func (h *Handle) XfrmPolicyDel(policy *XfrmPolicy) error {
-	_, err := h.xfrmPolicyGetOrDelete(policy, nl.XFRM_MSG_DELPOLICY)
+	_, err := h.xfrmPolicyGetOrDelete(policy, int(nl.XFRM_MSG_DELPOLICY))
 	return err
 }
 
@@ -130,12 +130,12 @@ func XfrmPolicyList(family int) ([]XfrmPolicy, error) {
 // Equivalent to: `ip xfrm policy show`.
 // The list can be filtered by ip family.
 func (h *Handle) XfrmPolicyList(family int) ([]XfrmPolicy, error) {
-	req := h.newNetlinkRequest(nl.XFRM_MSG_GETPOLICY, unix.NLM_F_DUMP)
+	req := h.newNetlinkRequest(int(nl.XFRM_MSG_GETPOLICY), unix.NLM_F_DUMP)
 
 	msg := nl.NewIfInfomsg(family)
 	req.AddData(msg)
 
-	msgs, err := req.Execute(unix.NETLINK_XFRM, nl.XFRM_MSG_NEWPOLICY)
+	msgs, err := req.Execute(unix.NETLINK_XFRM, uint16(nl.XFRM_MSG_NEWPOLICY))
 	if err != nil {
 		return nil, err
 	}
@@ -162,7 +162,7 @@ func XfrmPolicyGet(policy *XfrmPolicy) (*XfrmPolicy, error) {
 // XfrmPolicyGet gets a the policy described by the index or selector, if found.
 // Equivalent to: `ip xfrm policy get { SELECTOR | index INDEX } dir DIR [ctx CTX ] [ mark MARK [ mask MASK ] ] [ ptype PTYPE ]`.
 func (h *Handle) XfrmPolicyGet(policy *XfrmPolicy) (*XfrmPolicy, error) {
-	return h.xfrmPolicyGetOrDelete(policy, nl.XFRM_MSG_GETPOLICY)
+	return h.xfrmPolicyGetOrDelete(policy, int(nl.XFRM_MSG_GETPOLICY))
 }
 
 // XfrmPolicyFlush will flush the policies on the system.
@@ -174,7 +174,7 @@ func XfrmPolicyFlush() error {
 // XfrmPolicyFlush will flush the policies on the system.
 // Equivalent to: `ip xfrm policy flush`
 func (h *Handle) XfrmPolicyFlush() error {
-	req := h.newNetlinkRequest(nl.XFRM_MSG_FLUSHPOLICY, unix.NLM_F_ACK)
+	req := h.newNetlinkRequest(int(nl.XFRM_MSG_FLUSHPOLICY), unix.NLM_F_ACK)
 	_, err := req.Execute(unix.NETLINK_XFRM, 0)
 	return err
 }
@@ -199,7 +199,7 @@ func (h *Handle) xfrmPolicyGetOrDelete(policy *XfrmPolicy, nlProto int) (*XfrmPo
 	}
 
 	resType := nl.XFRM_MSG_NEWPOLICY
-	if nlProto == nl.XFRM_MSG_DELPOLICY {
+	if nlProto == int(nl.XFRM_MSG_DELPOLICY) {
 		resType = 0
 	}
 
@@ -208,7 +208,7 @@ func (h *Handle) xfrmPolicyGetOrDelete(policy *XfrmPolicy, nlProto int) (*XfrmPo
 		return nil, err
 	}
 
-	if nlProto == nl.XFRM_MSG_DELPOLICY {
+	if nlProto == int(nl.XFRM_MSG_DELPOLICY) {
 		return nil, err
 	}
